@@ -1,7 +1,7 @@
 {
   "name": "fresh",
   "description": "HTTP response freshness testing",
-  "version": "0.5.0",
+  "version": "0.5.2",
   "author": "TJ Holowaychuk <tj@vision-media.ca> (http://tjholowaychuk.com)",
   "contributors": [
     "Douglas Christopher Wilson <doug@somethingdoug.com>",
@@ -16,10 +16,15 @@
   ],
   "repository": "jshttp/fresh",
   "devDependencies": {
-    "eslint": "3.16.0",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-promise": "3.4.2",
-    "eslint-plugin-standard": "2.0.1",
+    "beautify-benchmark": "0.2.4",
+    "benchmark": "2.1.4",
+    "eslint": "3.19.0",
+    "eslint-config-standard": "10.2.1",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-markdown": "1.0.0-beta.6",
+    "eslint-plugin-node": "5.1.1",
+    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "1.21.5"
   },
@@ -32,7 +37,8 @@
     "node": ">= 0.6"
   },
   "scripts": {
-    "lint": "eslint .",
+    "bench": "node benchmark/index.js",
+    "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
