@@ -1,7 +1,7 @@
 {
   "name": "proxy-addr",
   "description": "Determine address of proxied request",
-  "version": "1.1.4",
+  "version": "2.0.2",
   "author": "Douglas Christopher Wilson <doug@somethingdoug.com>",
   "license": "MIT",
   "keywords": [
@@ -11,14 +11,21 @@
   ],
   "repository": "jshttp/proxy-addr",
   "dependencies": {
-    "forwarded": "~0.1.0",
-    "ipaddr.js": "1.3.0"
+    "forwarded": "~0.1.2",
+    "ipaddr.js": "1.5.2"
   },
   "devDependencies": {
-    "benchmark": "2.1.3",
+    "benchmark": "2.1.4",
     "beautify-benchmark": "0.2.4",
-    "istanbul": "0.4.5",
-    "mocha": "~1.21.5"
+    "eslint": "3.19.0",
+    "eslint-config-standard": "10.2.1",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-markdown": "1.0.0-beta.6",
+    "eslint-plugin-node": "5.1.1",
+    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-standard": "3.0.1",
+    "mocha": "3.5.3",
+    "nyc": "10.3.2"
   },
   "files": [
     "LICENSE",
@@ -27,12 +34,13 @@
     "index.js"
   ],
   "engines": {
-    "node": ">= 0.6"
+    "node": ">= 0.10"
   },
   "scripts": {
     "bench": "node benchmark/index.js",
+    "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
-    "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
-    "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/"
+    "test-cov": "nyc --reporter=text npm test",
+    "test-travis": "nyc --reporter=html --reporter=text npm test"
   }
 }
