@@ -1,7 +1,7 @@
 {
   "name": "etag",
   "description": "Create simple HTTP ETags",
-  "version": "1.8.0",
+  "version": "1.8.1",
   "contributors": [
     "Douglas Christopher Wilson <doug@somethingdoug.com>",
     "David Björklund <david.bjorklund@gmail.com>"
@@ -15,15 +15,18 @@
   "repository": "jshttp/etag",
   "devDependencies": {
     "beautify-benchmark": "0.2.4",
-    "benchmark": "2.1.3",
-    "eslint": "3.15.0",
-    "eslint-config-standard": "6.2.1",
-    "eslint-plugin-markdown": "1.0.0-beta.3",
-    "eslint-plugin-promise": "3.4.2",
-    "eslint-plugin-standard": "2.0.1",
+    "benchmark": "2.1.4",
+    "eslint": "3.19.0",
+    "eslint-config-standard": "10.2.1",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-markdown": "1.0.0-beta.6",
+    "eslint-plugin-node": "5.1.1",
+    "eslint-plugin-promise": "3.5.0",
+    "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "1.21.5",
-    "seedrandom": "2.4.2"
+    "safe-buffer": "5.1.1",
+    "seedrandom": "2.4.3"
   },
   "files": [
     "LICENSE",
