@@ -1,6 +1,6 @@
 /*!
  * finalhandler
- * Copyright(c) 2014-2017 Douglas Christopher Wilson
+ * Copyright(c) 2014-2015 Douglas Christopher Wilson
  * MIT Licensed
  */
 
@@ -12,11 +12,9 @@
  */
 
 var debug = require('debug')('finalhandler')
-var encodeUrl = require('encodeurl')
 var escapeHtml = require('escape-html')
+var http = require('http')
 var onFinished = require('on-finished')
-var parseUrl = require('parseurl')
-var statuses = require('statuses')
 var unpipe = require('unpipe')
 
 /**
@@ -24,38 +22,11 @@ var unpipe = require('unpipe')
  * @private
  */
 
-var DOUBLE_SPACE_REGEXP = /\x20{2}/g
-var NEWLINE_REGEXP = /\n/g
-
 /* istanbul ignore next */
 var defer = typeof setImmediate === 'function'
   ? setImmediate
-  : function (fn) { process.nextTick(fn.bind.apply(fn, arguments)) }
+  : function(fn){ process.nextTick(fn.bind.apply(fn, arguments)) }
 var isFinished = onFinished.isFinished
-
-/**
- * Create a minimal HTML document.
- *
- * @param {string} message
- * @private
- */
-
-function createHtmlDocument (message) {
-  var body = escapeHtml(message)
-    .replace(NEWLINE_REGEXP, '<br>')
-    .replace(DOUBLE_SPACE_REGEXP, ' &nbsp;')
-
-  return '<!DOCTYPE html>\n' +
-    '<html lang="en">\n' +
-    '<head>\n' +
-    '<meta charset="utf-8">\n' +
-    '<title>Error</title>\n' +
-    '</head>\n' +
-    '<body>\n' +
-    '<pre>' + body + '</pre>\n' +
-    '</body>\n' +
-    '</html>\n'
-}
 
 /**
  * Module exports.
@@ -74,7 +45,7 @@ module.exports = finalhandler
  * @public
  */
 
-function finalhandler (req, res, options) {
+function finalhandler(req, res, options) {
   var opts = options || {}
 
   // get environment
@@ -84,9 +55,7 @@ function finalhandler (req, res, options) {
   var onerror = opts.onerror
 
   return function (err) {
-    var headers
-    var msg
-    var status
+    var status = res.statusCode
 
     // ignore 404 on in-flight response
     if (!err && res._header) {
@@ -96,25 +65,31 @@ function finalhandler (req, res, options) {
 
     // unhandled error
     if (err) {
-      // respect status code from error
-      status = getErrorStatusCode(err)
-
-      // respect headers from error
-      if (status !== undefined) {
-        headers = getErrorHeaders(err)
+      // respect err.statusCode
+      if (err.statusCode) {
+        status = err.statusCode
       }
 
-      // fallback to status code on response
-      if (status === undefined) {
-        status = getResponseStatusCode(res)
+      // respect err.status
+      if (err.status) {
+        status = err.status
       }
 
-      // get error message
-      msg = getErrorMessage(err, status, env)
+      // default status code to 500
+      if (!status || status < 400) {
+        status = 500
+      }
+
+      // production gets a basic error message
+      var msg = env === 'production'
+        ? http.STATUS_CODES[status]
+        : err.stack || err.toString()
+      msg = escapeHtml(msg)
+        .replace(/\n/g, '<br>')
+        .replace(/  /g, ' &nbsp;') + '\n'
     } else {
-      // not found
       status = 404
-      msg = 'Cannot ' + req.method + ' ' + encodeUrl(parseUrl.original(req).pathname)
+      msg = 'Cannot ' + escapeHtml(req.method) + ' ' + escapeHtml(req.originalUrl || req.url) + '\n'
     }
 
     debug('default %s', status)
@@ -126,105 +101,11 @@ function finalhandler (req, res, options) {
 
     // cannot actually respond
     if (res._header) {
-      debug('cannot %d after headers sent', status)
-      req.socket.destroy()
-      return
+      return req.socket.destroy()
     }
 
-    // send response
-    send(req, res, status, headers, msg)
+    send(req, res, status, msg)
   }
-}
-
-/**
- * Get headers from Error object.
- *
- * @param {Error} err
- * @return {object}
- * @private
- */
-
-function getErrorHeaders (err) {
-  if (!err.headers || typeof err.headers !== 'object') {
-    return undefined
-  }
-
-  var headers = Object.create(null)
-  var keys = Object.keys(err.headers)
-
-  for (var i = 0; i < keys.length; i++) {
-    var key = keys[i]
-    headers[key] = err.headers[key]
-  }
-
-  return headers
-}
-
-/**
- * Get message from Error object, fallback to status message.
- *
- * @param {Error} err
- * @param {number} status
- * @param {string} env
- * @return {string}
- * @private
- */
-
-function getErrorMessage (err, status, env) {
-  var msg
-
-  if (env !== 'production') {
-    // use err.stack, which typically includes err.message
-    msg = err.stack
-
-    // fallback to err.toString() when possible
-    if (!msg && typeof err.toString === 'function') {
-      msg = err.toString()
-    }
-  }
-
-  return msg || statuses[status]
-}
-
-/**
- * Get status code from Error object.
- *
- * @param {Error} err
- * @return {number}
- * @private
- */
-
-function getErrorStatusCode (err) {
-  // check err.status
-  if (typeof err.status === 'number' && err.status >= 400 && err.status < 600) {
-    return err.status
-  }
-
-  // check err.statusCode
-  if (typeof err.statusCode === 'number' && err.statusCode >= 400 && err.statusCode < 600) {
-    return err.statusCode
-  }
-
-  return undefined
-}
-
-/**
- * Get status code from response.
- *
- * @param {OutgoingMessage} res
- * @return {number}
- * @private
- */
-
-function getResponseStatusCode (res) {
-  var status = res.statusCode
-
-  // default status code to 500 if outside valid range
-  if (typeof status !== 'number' || status < 400 || status > 599) {
-    status = 500
-  }
-
-  return status
 }
 
 /**
@@ -233,25 +114,15 @@ function getResponseStatusCode (res) {
  * @param {IncomingMessage} req
  * @param {OutgoingMessage} res
  * @param {number} status
- * @param {object} headers
- * @param {string} message
+ * @param {string} body
  * @private
  */
 
-function send (req, res, status, headers, message) {
-  function write () {
-    // response body
-    var body = createHtmlDocument(message)
-
-    // response status
+function send(req, res, status, body) {
+  function write() {
     res.statusCode = status
-    res.statusMessage = statuses[status]
 
-    // response headers
-    setHeaders(res, headers)
-
-    // security headers
-    res.setHeader('Content-Security-Policy', "default-src 'self'")
+    // security header for content sniffing
     res.setHeader('X-Content-Type-Options', 'nosniff')
 
     // standard headers
@@ -277,24 +148,4 @@ function send (req, res, status, headers, message) {
   // flush the request
   onFinished(req, write)
   req.resume()
-}
-
-/**
- * Set response headers from an object.
- *
- * @param {OutgoingMessage} res
- * @param {object} headers
- * @private
- */
-
-function setHeaders (res, headers) {
-  if (!headers) {
-    return
-  }
-
-  var keys = Object.keys(headers)
-  for (var i = 0; i < keys.length; i++) {
-    var key = keys[i]
-    res.setHeader(key, headers[key])
-  }
 }
