@@ -1,5 +1,5 @@
 (function() {
-  var expandIPv6, ipaddr, ipv4Part, ipv4Regexes, ipv6Part, ipv6Regexes, matchCIDR, root;
+  var expandIPv6, ipaddr, ipv4Part, ipv4Regexes, ipv6Part, ipv6Regexes, matchCIDR, root, zoneIndex;
 
   ipaddr = {};
 
@@ -32,7 +32,7 @@
   };
 
   ipaddr.subnetMatch = function(address, rangeList, defaultName) {
-    var rangeName, rangeSubnets, subnet, _i, _len;
+    var k, len, rangeName, rangeSubnets, subnet;
     if (defaultName == null) {
       defaultName = 'unicast';
     }
@@ -41,10 +41,12 @@
       if (rangeSubnets[0] && !(rangeSubnets[0] instanceof Array)) {
         rangeSubnets = [rangeSubnets];
       }
-      for (_i = 0, _len = rangeSubnets.length; _i < _len; _i++) {
-        subnet = rangeSubnets[_i];
-        if (address.match.apply(address, subnet)) {
-          return rangeName;
+      for (k = 0, len = rangeSubnets.length; k < len; k++) {
+        subnet = rangeSubnets[k];
+        if (address.kind() === subnet[0].kind()) {
+          if (address.match.apply(address, subnet)) {
+            return rangeName;
+          }
         }
       }
     }
@@ -53,12 +55,12 @@
 
   ipaddr.IPv4 = (function() {
     function IPv4(octets) {
-      var octet, _i, _len;
+      var k, len, octet;
       if (octets.length !== 4) {
         throw new Error("ipaddr: ipv4 octet count should be 4");
       }
-      for (_i = 0, _len = octets.length; _i < _len; _i++) {
-        octet = octets[_i];
+      for (k = 0, len = octets.length; k < len; k++) {
+        octet = octets[k];
         if (!((0 <= octet && octet <= 255))) {
           throw new Error("ipaddr: ipv4 octet should fit in 8 bits");
         }
@@ -74,14 +76,18 @@
       return this.octets.join(".");
     };
 
+    IPv4.prototype.toNormalizedString = function() {
+      return this.toString();
+    };
+
     IPv4.prototype.toByteArray = function() {
       return this.octets.slice(0);
     };
 
     IPv4.prototype.match = function(other, cidrRange) {
-      var _ref;
+      var ref;
       if (cidrRange === void 0) {
-        _ref = other, other = _ref[0], cidrRange = _ref[1];
+        ref = other, other = ref[0], cidrRange = ref[1];
       }
       if (other.kind() !== 'ipv4') {
         throw new Error("ipaddr: cannot match ipv4 address with non-ipv4 one");
@@ -109,7 +115,7 @@
     };
 
     IPv4.prototype.prefixLengthFromSubnetMask = function() {
-      var cidr, i, octet, stop, zeros, zerotable, _i;
+      var cidr, i, k, octet, stop, zeros, zerotable;
       zerotable = {
         0: 8,
         128: 7,
@@ -123,7 +129,7 @@
       };
       cidr = 0;
       stop = false;
-      for (i = _i = 3; _i >= 0; i = _i += -1) {
+      for (i = k = 3; k >= 0; i = k += -1) {
         octet = this.octets[i];
         if (octet in zerotable) {
           zeros = zerotable[octet];
@@ -163,14 +169,14 @@
     };
     if (match = string.match(ipv4Regexes.fourOctet)) {
       return (function() {
-        var _i, _len, _ref, _results;
-        _ref = match.slice(1, 6);
-        _results = [];
-        for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-          part = _ref[_i];
-          _results.push(parseIntAuto(part));
+        var k, len, ref, results;
+        ref = match.slice(1, 6);
+        results = [];
+        for (k = 0, len = ref.length; k < len; k++) {
+          part = ref[k];
+          results.push(parseIntAuto(part));
         }
-        return _results;
+        return results;
       })();
     } else if (match = string.match(ipv4Regexes.longValue)) {
       value = parseIntAuto(match[1]);
@@ -178,12 +184,12 @@
         throw new Error("ipaddr: address outside defined range");
       }
       return ((function() {
-        var _i, _results;
-        _results = [];
-        for (shift = _i = 0; _i <= 24; shift = _i += 8) {
-          _results.push((value >> shift) & 0xff);
+        var k, results;
+        results = [];
+        for (shift = k = 0; k <= 24; shift = k += 8) {
+          results.push((value >> shift) & 0xff);
         }
-        return _results;
+        return results;
       })()).reverse();
     } else {
       return null;
@@ -191,11 +197,11 @@
   };
 
   ipaddr.IPv6 = (function() {
-    function IPv6(parts) {
-      var i, part, _i, _j, _len, _ref;
+    function IPv6(parts, zoneId) {
+      var i, k, l, len, part, ref;
       if (parts.length === 16) {
         this.parts = [];
-        for (i = _i = 0; _i <= 14; i = _i += 2) {
+        for (i = k = 0; k <= 14; i = k += 2) {
           this.parts.push((parts[i] << 8) | parts[i + 1]);
         }
       } else if (parts.length === 8) {
@@ -203,12 +209,15 @@
       } else {
         throw new Error("ipaddr: ipv6 part count should be 8 or 16");
       }
-      _ref = this.parts;
-      for (_j = 0, _len = _ref.length; _j < _len; _j++) {
-        part = _ref[_j];
+      ref = this.parts;
+      for (l = 0, len = ref.length; l < len; l++) {
+        part = ref[l];
         if (!((0 <= part && part <= 0xffff))) {
           throw new Error("ipaddr: ipv6 part should fit in 16 bits");
         }
+      }
+      if (zoneId) {
+        this.zoneId = zoneId;
       }
     }
 
@@ -217,24 +226,24 @@
     };
 
     IPv6.prototype.toString = function() {
-      var compactStringParts, part, pushPart, state, stringParts, _i, _len;
+      var addr, compactStringParts, k, len, part, pushPart, state, stringParts, suffix;
       stringParts = (function() {
-        var _i, _len, _ref, _results;
-        _ref = this.parts;
-        _results = [];
-        for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-          part = _ref[_i];
-          _results.push(part.toString(16));
+        var k, len, ref, results;
+        ref = this.parts;
+        results = [];
+        for (k = 0, len = ref.length; k < len; k++) {
+          part = ref[k];
+          results.push(part.toString(16));
         }
-        return _results;
+        return results;
       }).call(this);
       compactStringParts = [];
       pushPart = function(part) {
         return compactStringParts.push(part);
       };
       state = 0;
-      for (_i = 0, _len = stringParts.length; _i < _len; _i++) {
-        part = stringParts[_i];
+      for (k = 0, len = stringParts.length; k < len; k++) {
+        part = stringParts[k];
         switch (state) {
           case 0:
             if (part === '0') {
@@ -266,15 +275,20 @@
         pushPart('');
         pushPart('');
       }
-      return compactStringParts.join(":");
+      addr = compactStringParts.join(":");
+      suffix = '';
+      if (this.zoneId) {
+        suffix = '%' + this.zoneId;
+      }
+      return addr + suffix;
     };
 
     IPv6.prototype.toByteArray = function() {
-      var bytes, part, _i, _len, _ref;
+      var bytes, k, len, part, ref;
       bytes = [];
-      _ref = this.parts;
-      for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-        part = _ref[_i];
+      ref = this.parts;
+      for (k = 0, len = ref.length; k < len; k++) {
+        part = ref[k];
         bytes.push(part >> 8);
         bytes.push(part & 0xff);
       }
@@ -282,23 +296,28 @@
     };
 
     IPv6.prototype.toNormalizedString = function() {
-      var part;
-      return ((function() {
-        var _i, _len, _ref, _results;
-        _ref = this.parts;
-        _results = [];
-        for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-          part = _ref[_i];
-          _results.push(part.toString(16));
+      var addr, part, suffix;
+      addr = ((function() {
+        var k, len, ref, results;
+        ref = this.parts;
+        results = [];
+        for (k = 0, len = ref.length; k < len; k++) {
+          part = ref[k];
+          results.push(part.toString(16));
         }
-        return _results;
+        return results;
       }).call(this)).join(":");
+      suffix = '';
+      if (this.zoneId) {
+        suffix = '%' + this.zoneId;
+      }
+      return addr + suffix;
     };
 
     IPv6.prototype.match = function(other, cidrRange) {
-      var _ref;
+      var ref;
       if (cidrRange === void 0) {
-        _ref = other, other = _ref[0], cidrRange = _ref[1];
+        ref = other, other = ref[0], cidrRange = ref[1];
       }
       if (other.kind() !== 'ipv6') {
         throw new Error("ipaddr: cannot match ipv6 address with non-ipv6 one");
@@ -329,12 +348,53 @@
     };
 
     IPv6.prototype.toIPv4Address = function() {
-      var high, low, _ref;
+      var high, low, ref;
       if (!this.isIPv4MappedAddress()) {
         throw new Error("ipaddr: trying to convert a generic ipv6 address to ipv4");
       }
-      _ref = this.parts.slice(-2), high = _ref[0], low = _ref[1];
+      ref = this.parts.slice(-2), high = ref[0], low = ref[1];
       return new ipaddr.IPv4([high >> 8, high & 0xff, low >> 8, low & 0xff]);
+    };
+
+    IPv6.prototype.prefixLengthFromSubnetMask = function() {
+      var cidr, i, k, part, stop, zeros, zerotable;
+      zerotable = {
+        0: 16,
+        32768: 15,
+        49152: 14,
+        57344: 13,
+        61440: 12,
+        63488: 11,
+        64512: 10,
+        65024: 9,
+        65280: 8,
+        65408: 7,
+        65472: 6,
+        65504: 5,
+        65520: 4,
+        65528: 3,
+        65532: 2,
+        65534: 1,
+        65535: 0
+      };
+      cidr = 0;
+      stop = false;
+      for (i = k = 7; k >= 0; i = k += -1) {
+        part = this.parts[i];
+        if (part in zerotable) {
+          zeros = zerotable[part];
+          if (stop && zeros !== 0) {
+            return null;
+          }
+          if (zeros !== 16) {
+            stop = true;
+          }
+          cidr += zeros;
+        } else {
+          return null;
+        }
+      }
+      return 128 - cidr;
     };
 
     return IPv6;
@@ -343,15 +403,23 @@
 
   ipv6Part = "(?:[0-9a-f]+::?)+";
 
+  zoneIndex = "%[0-9a-z]{1,}";
+
   ipv6Regexes = {
-    "native": new RegExp("^(::)?(" + ipv6Part + ")?([0-9a-f]+)?(::)?$", 'i'),
-    transitional: new RegExp(("^((?:" + ipv6Part + ")|(?:::)(?:" + ipv6Part + ")?)") + ("" + ipv4Part + "\\." + ipv4Part + "\\." + ipv4Part + "\\." + ipv4Part + "$"), 'i')
+    zoneIndex: new RegExp(zoneIndex, 'i'),
+    "native": new RegExp("^(::)?(" + ipv6Part + ")?([0-9a-f]+)?(::)?(" + zoneIndex + ")?$", 'i'),
+    transitional: new RegExp(("^((?:" + ipv6Part + ")|(?:::)(?:" + ipv6Part + ")?)") + (ipv4Part + "\\." + ipv4Part + "\\." + ipv4Part + "\\." + ipv4Part) + ("(" + zoneIndex + ")?$"), 'i')
   };
 
   expandIPv6 = function(string, parts) {
-    var colonCount, lastColon, part, replacement, replacementCount;
+    var colonCount, lastColon, part, replacement, replacementCount, zoneId;
     if (string.indexOf('::') !== string.lastIndexOf('::')) {
       return null;
+    }
+    zoneId = (string.match(ipv6Regexes['zoneIndex']) || [])[0];
+    if (zoneId) {
+      zoneId = zoneId.substring(1);
+      string = string.replace(/%.+$/, '');
     }
     colonCount = 0;
     lastColon = -1;
@@ -379,35 +447,43 @@
     if (string[string.length - 1] === ':') {
       string = string.slice(0, -1);
     }
-    return (function() {
-      var _i, _len, _ref, _results;
-      _ref = string.split(":");
-      _results = [];
-      for (_i = 0, _len = _ref.length; _i < _len; _i++) {
-        part = _ref[_i];
-        _results.push(parseInt(part, 16));
+    parts = (function() {
+      var k, len, ref, results;
+      ref = string.split(":");
+      results = [];
+      for (k = 0, len = ref.length; k < len; k++) {
+        part = ref[k];
+        results.push(parseInt(part, 16));
       }
-      return _results;
+      return results;
     })();
+    return {
+      parts: parts,
+      zoneId: zoneId
+    };
   };
 
   ipaddr.IPv6.parser = function(string) {
-    var match, octet, octets, parts, _i, _len;
-    if (string.match(ipv6Regexes['native'])) {
+    var addr, k, len, match, octet, octets, zoneId;
+    if (ipv6Regexes['native'].test(string)) {
       return expandIPv6(string, 8);
     } else if (match = string.match(ipv6Regexes['transitional'])) {
-      parts = expandIPv6(match[1].slice(0, -1), 6);
-      if (parts) {
+      zoneId = match[6] || '';
+      addr = expandIPv6(match[1].slice(0, -1) + zoneId, 6);
+      if (addr.parts) {
         octets = [parseInt(match[2]), parseInt(match[3]), parseInt(match[4]), parseInt(match[5])];
-        for (_i = 0, _len = octets.length; _i < _len; _i++) {
-          octet = octets[_i];
+        for (k = 0, len = octets.length; k < len; k++) {
+          octet = octets[k];
           if (!((0 <= octet && octet <= 255))) {
             return null;
           }
         }
-        parts.push(octets[0] << 8 | octets[1]);
-        parts.push(octets[2] << 8 | octets[3]);
-        return parts;
+        addr.parts.push(octets[0] << 8 | octets[1]);
+        addr.parts.push(octets[2] << 8 | octets[3]);
+        return {
+          parts: addr.parts,
+          zoneId: addr.zoneId
+        };
       }
     }
     return null;
@@ -422,8 +498,8 @@
     try {
       new this(this.parser(string));
       return true;
-    } catch (_error) {
-      e = _error;
+    } catch (error1) {
+      e = error1;
       return false;
     }
   };
@@ -437,26 +513,36 @@
   };
 
   ipaddr.IPv6.isValid = function(string) {
-    var e;
+    var addr, e;
     if (typeof string === "string" && string.indexOf(":") === -1) {
       return false;
     }
     try {
-      new this(this.parser(string));
+      addr = this.parser(string);
+      new this(addr.parts, addr.zoneId);
       return true;
-    } catch (_error) {
-      e = _error;
+    } catch (error1) {
+      e = error1;
       return false;
     }
   };
 
-  ipaddr.IPv4.parse = ipaddr.IPv6.parse = function(string) {
+  ipaddr.IPv4.parse = function(string) {
     var parts;
     parts = this.parser(string);
     if (parts === null) {
       throw new Error("ipaddr: string is not formatted like ip address");
     }
     return new this(parts);
+  };
+
+  ipaddr.IPv6.parse = function(string) {
+    var addr;
+    addr = this.parser(string);
+    if (addr.parts === null) {
+      throw new Error("ipaddr: string is not formatted like ip address");
+    }
+    return new this(addr.parts, addr.zoneId);
   };
 
   ipaddr.IPv4.parseCIDR = function(string) {
@@ -468,6 +554,63 @@
       }
     }
     throw new Error("ipaddr: string is not formatted like an IPv4 CIDR range");
+  };
+
+  ipaddr.IPv4.subnetMaskFromPrefixLength = function(prefix) {
+    var filledOctetCount, j, octets;
+    prefix = parseInt(prefix);
+    if (prefix < 0 || prefix > 32) {
+      throw new Error('ipaddr: invalid IPv4 prefix length');
+    }
+    octets = [0, 0, 0, 0];
+    j = 0;
+    filledOctetCount = Math.floor(prefix / 8);
+    while (j < filledOctetCount) {
+      octets[j] = 255;
+      j++;
+    }
+    if (filledOctetCount < 4) {
+      octets[filledOctetCount] = Math.pow(2, prefix % 8) - 1 << 8 - (prefix % 8);
+    }
+    return new this(octets);
+  };
+
+  ipaddr.IPv4.broadcastAddressFromCIDR = function(string) {
+    var cidr, error, i, ipInterfaceOctets, octets, subnetMaskOctets;
+    try {
+      cidr = this.parseCIDR(string);
+      ipInterfaceOctets = cidr[0].toByteArray();
+      subnetMaskOctets = this.subnetMaskFromPrefixLength(cidr[1]).toByteArray();
+      octets = [];
+      i = 0;
+      while (i < 4) {
+        octets.push(parseInt(ipInterfaceOctets[i], 10) | parseInt(subnetMaskOctets[i], 10) ^ 255);
+        i++;
+      }
+      return new this(octets);
+    } catch (error1) {
+      error = error1;
+      throw new Error('ipaddr: the address does not have IPv4 CIDR format');
+    }
+  };
+
+  ipaddr.IPv4.networkAddressFromCIDR = function(string) {
+    var cidr, error, i, ipInterfaceOctets, octets, subnetMaskOctets;
+    try {
+      cidr = this.parseCIDR(string);
+      ipInterfaceOctets = cidr[0].toByteArray();
+      subnetMaskOctets = this.subnetMaskFromPrefixLength(cidr[1]).toByteArray();
+      octets = [];
+      i = 0;
+      while (i < 4) {
+        octets.push(parseInt(ipInterfaceOctets[i], 10) & parseInt(subnetMaskOctets[i], 10));
+        i++;
+      }
+      return new this(octets);
+    } catch (error1) {
+      error = error1;
+      throw new Error('ipaddr: the address does not have IPv4 CIDR format');
+    }
   };
 
   ipaddr.IPv6.parseCIDR = function(string) {
@@ -499,12 +642,12 @@
     var e;
     try {
       return ipaddr.IPv6.parseCIDR(string);
-    } catch (_error) {
-      e = _error;
+    } catch (error1) {
+      e = error1;
       try {
         return ipaddr.IPv4.parseCIDR(string);
-      } catch (_error) {
-        e = _error;
+      } catch (error1) {
+        e = error1;
         throw new Error("ipaddr: the address has neither IPv6 nor IPv4 CIDR format");
       }
     }
