@@ -1,519 +1,478 @@
-'use strict';
+/* eslint no-extend-native:0 */
+// Load modules
 
-var test = require('tape');
-var qs = require('../');
-var iconv = require('iconv-lite');
+var Code = require('code');
+var Lab = require('lab');
+var Qs = require('../');
 
-test('parse()', function (t) {
-    t.test('parses a simple string', function (st) {
-        st.deepEqual(qs.parse('0=foo'), { 0: 'foo' });
-        st.deepEqual(qs.parse('foo=c++'), { foo: 'c  ' });
-        st.deepEqual(qs.parse('a[>=]=23'), { a: { '>=': '23' } });
-        st.deepEqual(qs.parse('a[<=>]==23'), { a: { '<=>': '=23' } });
-        st.deepEqual(qs.parse('a[==]=23'), { a: { '==': '23' } });
-        st.deepEqual(qs.parse('foo', { strictNullHandling: true }), { foo: null });
-        st.deepEqual(qs.parse('foo'), { foo: '' });
-        st.deepEqual(qs.parse('foo='), { foo: '' });
-        st.deepEqual(qs.parse('foo=bar'), { foo: 'bar' });
-        st.deepEqual(qs.parse(' foo = bar = baz '), { ' foo ': ' bar = baz ' });
-        st.deepEqual(qs.parse('foo=bar=baz'), { foo: 'bar=baz' });
-        st.deepEqual(qs.parse('foo=bar&bar=baz'), { foo: 'bar', bar: 'baz' });
-        st.deepEqual(qs.parse('foo2=bar2&baz2='), { foo2: 'bar2', baz2: '' });
-        st.deepEqual(qs.parse('foo=bar&baz', { strictNullHandling: true }), { foo: 'bar', baz: null });
-        st.deepEqual(qs.parse('foo=bar&baz'), { foo: 'bar', baz: '' });
-        st.deepEqual(qs.parse('cht=p3&chd=t:60,40&chs=250x100&chl=Hello|World'), {
+
+// Declare internals
+
+var internals = {};
+
+
+// Test shortcuts
+
+var lab = exports.lab = Lab.script();
+var expect = Code.expect;
+var describe = lab.experiment;
+var it = lab.test;
+
+
+describe('parse()', function () {
+
+    it('parses a simple string', function (done) {
+
+        expect(Qs.parse('0=foo')).to.deep.equal({ '0': 'foo' });
+        expect(Qs.parse('foo=c++')).to.deep.equal({ foo: 'c  ' });
+        expect(Qs.parse('a[>=]=23')).to.deep.equal({ a: { '>=': '23' } });
+        expect(Qs.parse('a[<=>]==23')).to.deep.equal({ a: { '<=>': '=23' } });
+        expect(Qs.parse('a[==]=23')).to.deep.equal({ a: { '==': '23' } });
+        expect(Qs.parse('foo', { strictNullHandling: true })).to.deep.equal({ foo: null });
+        expect(Qs.parse('foo' )).to.deep.equal({ foo: '' });
+        expect(Qs.parse('foo=')).to.deep.equal({ foo: '' });
+        expect(Qs.parse('foo=bar')).to.deep.equal({ foo: 'bar' });
+        expect(Qs.parse(' foo = bar = baz ')).to.deep.equal({ ' foo ': ' bar = baz ' });
+        expect(Qs.parse('foo=bar=baz')).to.deep.equal({ foo: 'bar=baz' });
+        expect(Qs.parse('foo=bar&bar=baz')).to.deep.equal({ foo: 'bar', bar: 'baz' });
+        expect(Qs.parse('foo2=bar2&baz2=')).to.deep.equal({ foo2: 'bar2', baz2: '' });
+        expect(Qs.parse('foo=bar&baz', { strictNullHandling: true })).to.deep.equal({ foo: 'bar', baz: null });
+        expect(Qs.parse('foo=bar&baz')).to.deep.equal({ foo: 'bar', baz: '' });
+        expect(Qs.parse('cht=p3&chd=t:60,40&chs=250x100&chl=Hello|World')).to.deep.equal({
             cht: 'p3',
             chd: 't:60,40',
             chs: '250x100',
             chl: 'Hello|World'
         });
-        st.end();
+        done();
     });
 
-    t.test('allows enabling dot notation', function (st) {
-        st.deepEqual(qs.parse('a.b=c'), { 'a.b': 'c' });
-        st.deepEqual(qs.parse('a.b=c', { allowDots: true }), { a: { b: 'c' } });
-        st.end();
+    it('allows disabling dot notation', function (done) {
+
+        expect(Qs.parse('a.b=c')).to.deep.equal({ a: { b: 'c' } });
+        expect(Qs.parse('a.b=c', { allowDots: false })).to.deep.equal({ 'a.b': 'c' });
+        done();
     });
 
-    t.deepEqual(qs.parse('a[b]=c'), { a: { b: 'c' } }, 'parses a single nested string');
-    t.deepEqual(qs.parse('a[b][c]=d'), { a: { b: { c: 'd' } } }, 'parses a double nested string');
-    t.deepEqual(
-        qs.parse('a[b][c][d][e][f][g][h]=i'),
-        { a: { b: { c: { d: { e: { f: { '[g][h]': 'i' } } } } } } },
-        'defaults to a depth of 5'
-    );
+    it('parses a single nested string', function (done) {
 
-    t.test('only parses one level when depth = 1', function (st) {
-        st.deepEqual(qs.parse('a[b][c]=d', { depth: 1 }), { a: { b: { '[c]': 'd' } } });
-        st.deepEqual(qs.parse('a[b][c][d]=e', { depth: 1 }), { a: { b: { '[c][d]': 'e' } } });
-        st.end();
+        expect(Qs.parse('a[b]=c')).to.deep.equal({ a: { b: 'c' } });
+        done();
     });
 
-    t.deepEqual(qs.parse('a=b&a=c'), { a: ['b', 'c'] }, 'parses a simple array');
+    it('parses a double nested string', function (done) {
 
-    t.test('parses an explicit array', function (st) {
-        st.deepEqual(qs.parse('a[]=b'), { a: ['b'] });
-        st.deepEqual(qs.parse('a[]=b&a[]=c'), { a: ['b', 'c'] });
-        st.deepEqual(qs.parse('a[]=b&a[]=c&a[]=d'), { a: ['b', 'c', 'd'] });
-        st.end();
+        expect(Qs.parse('a[b][c]=d')).to.deep.equal({ a: { b: { c: 'd' } } });
+        done();
     });
 
-    t.test('parses a mix of simple and explicit arrays', function (st) {
-        st.deepEqual(qs.parse('a=b&a[]=c'), { a: ['b', 'c'] });
-        st.deepEqual(qs.parse('a[]=b&a=c'), { a: ['b', 'c'] });
-        st.deepEqual(qs.parse('a[0]=b&a=c'), { a: ['b', 'c'] });
-        st.deepEqual(qs.parse('a=b&a[0]=c'), { a: ['b', 'c'] });
+    it('defaults to a depth of 5', function (done) {
 
-        st.deepEqual(qs.parse('a[1]=b&a=c', { arrayLimit: 20 }), { a: ['b', 'c'] });
-        st.deepEqual(qs.parse('a[]=b&a=c', { arrayLimit: 0 }), { a: ['b', 'c'] });
-        st.deepEqual(qs.parse('a[]=b&a=c'), { a: ['b', 'c'] });
-
-        st.deepEqual(qs.parse('a=b&a[1]=c', { arrayLimit: 20 }), { a: ['b', 'c'] });
-        st.deepEqual(qs.parse('a=b&a[]=c', { arrayLimit: 0 }), { a: ['b', 'c'] });
-        st.deepEqual(qs.parse('a=b&a[]=c'), { a: ['b', 'c'] });
-
-        st.end();
+        expect(Qs.parse('a[b][c][d][e][f][g][h]=i')).to.deep.equal({ a: { b: { c: { d: { e: { f: { '[g][h]': 'i' } } } } } } });
+        done();
     });
 
-    t.test('parses a nested array', function (st) {
-        st.deepEqual(qs.parse('a[b][]=c&a[b][]=d'), { a: { b: ['c', 'd'] } });
-        st.deepEqual(qs.parse('a[>=]=25'), { a: { '>=': '25' } });
-        st.end();
+    it('only parses one level when depth = 1', function (done) {
+
+        expect(Qs.parse('a[b][c]=d', { depth: 1 })).to.deep.equal({ a: { b: { '[c]': 'd' } } });
+        expect(Qs.parse('a[b][c][d]=e', { depth: 1 })).to.deep.equal({ a: { b: { '[c][d]': 'e' } } });
+        done();
     });
 
-    t.test('allows to specify array indices', function (st) {
-        st.deepEqual(qs.parse('a[1]=c&a[0]=b&a[2]=d'), { a: ['b', 'c', 'd'] });
-        st.deepEqual(qs.parse('a[1]=c&a[0]=b'), { a: ['b', 'c'] });
-        st.deepEqual(qs.parse('a[1]=c', { arrayLimit: 20 }), { a: ['c'] });
-        st.deepEqual(qs.parse('a[1]=c', { arrayLimit: 0 }), { a: { 1: 'c' } });
-        st.deepEqual(qs.parse('a[1]=c'), { a: ['c'] });
-        st.end();
+    it('parses a simple array', function (done) {
+
+        expect(Qs.parse('a=b&a=c')).to.deep.equal({ a: ['b', 'c'] });
+        done();
     });
 
-    t.test('limits specific array indices to arrayLimit', function (st) {
-        st.deepEqual(qs.parse('a[20]=a', { arrayLimit: 20 }), { a: ['a'] });
-        st.deepEqual(qs.parse('a[21]=a', { arrayLimit: 20 }), { a: { 21: 'a' } });
-        st.end();
+    it('parses an explicit array', function (done) {
+
+        expect(Qs.parse('a[]=b')).to.deep.equal({ a: ['b'] });
+        expect(Qs.parse('a[]=b&a[]=c')).to.deep.equal({ a: ['b', 'c'] });
+        expect(Qs.parse('a[]=b&a[]=c&a[]=d')).to.deep.equal({ a: ['b', 'c', 'd'] });
+        done();
     });
 
-    t.deepEqual(qs.parse('a[12b]=c'), { a: { '12b': 'c' } }, 'supports keys that begin with a number');
+    it('parses a mix of simple and explicit arrays', function (done) {
 
-    t.test('supports encoded = signs', function (st) {
-        st.deepEqual(qs.parse('he%3Dllo=th%3Dere'), { 'he=llo': 'th=ere' });
-        st.end();
+        expect(Qs.parse('a=b&a[]=c')).to.deep.equal({ a: ['b', 'c'] });
+        expect(Qs.parse('a[]=b&a=c')).to.deep.equal({ a: ['b', 'c'] });
+        expect(Qs.parse('a[0]=b&a=c')).to.deep.equal({ a: ['b', 'c'] });
+        expect(Qs.parse('a=b&a[0]=c')).to.deep.equal({ a: ['b', 'c'] });
+        expect(Qs.parse('a[1]=b&a=c')).to.deep.equal({ a: ['b', 'c'] });
+        expect(Qs.parse('a=b&a[1]=c')).to.deep.equal({ a: ['b', 'c'] });
+        done();
     });
 
-    t.test('is ok with url encoded strings', function (st) {
-        st.deepEqual(qs.parse('a[b%20c]=d'), { a: { 'b c': 'd' } });
-        st.deepEqual(qs.parse('a[b]=c%20d'), { a: { b: 'c d' } });
-        st.end();
+    it('parses a nested array', function (done) {
+
+        expect(Qs.parse('a[b][]=c&a[b][]=d')).to.deep.equal({ a: { b: ['c', 'd'] } });
+        expect(Qs.parse('a[>=]=25')).to.deep.equal({ a: { '>=': '25' } });
+        done();
     });
 
-    t.test('allows brackets in the value', function (st) {
-        st.deepEqual(qs.parse('pets=["tobi"]'), { pets: '["tobi"]' });
-        st.deepEqual(qs.parse('operators=[">=", "<="]'), { operators: '[">=", "<="]' });
-        st.end();
+    it('allows to specify array indices', function (done) {
+
+        expect(Qs.parse('a[1]=c&a[0]=b&a[2]=d')).to.deep.equal({ a: ['b', 'c', 'd'] });
+        expect(Qs.parse('a[1]=c&a[0]=b')).to.deep.equal({ a: ['b', 'c'] });
+        expect(Qs.parse('a[1]=c')).to.deep.equal({ a: ['c'] });
+        done();
     });
 
-    t.test('allows empty values', function (st) {
-        st.deepEqual(qs.parse(''), {});
-        st.deepEqual(qs.parse(null), {});
-        st.deepEqual(qs.parse(undefined), {});
-        st.end();
+    it('limits specific array indices to 20', function (done) {
+
+        expect(Qs.parse('a[20]=a')).to.deep.equal({ a: ['a'] });
+        expect(Qs.parse('a[21]=a')).to.deep.equal({ a: { '21': 'a' } });
+        done();
     });
 
-    t.test('transforms arrays to objects', function (st) {
-        st.deepEqual(qs.parse('foo[0]=bar&foo[bad]=baz'), { foo: { 0: 'bar', bad: 'baz' } });
-        st.deepEqual(qs.parse('foo[bad]=baz&foo[0]=bar'), { foo: { bad: 'baz', 0: 'bar' } });
-        st.deepEqual(qs.parse('foo[bad]=baz&foo[]=bar'), { foo: { bad: 'baz', 0: 'bar' } });
-        st.deepEqual(qs.parse('foo[]=bar&foo[bad]=baz'), { foo: { 0: 'bar', bad: 'baz' } });
-        st.deepEqual(qs.parse('foo[bad]=baz&foo[]=bar&foo[]=foo'), { foo: { bad: 'baz', 0: 'bar', 1: 'foo' } });
-        st.deepEqual(qs.parse('foo[0][a]=a&foo[0][b]=b&foo[1][a]=aa&foo[1][b]=bb'), { foo: [{ a: 'a', b: 'b' }, { a: 'aa', b: 'bb' }] });
+    it('supports keys that begin with a number', function (done) {
 
-        st.deepEqual(qs.parse('a[]=b&a[t]=u&a[hasOwnProperty]=c', { allowPrototypes: false }), { a: { 0: 'b', t: 'u' } });
-        st.deepEqual(qs.parse('a[]=b&a[t]=u&a[hasOwnProperty]=c', { allowPrototypes: true }), { a: { 0: 'b', t: 'u', hasOwnProperty: 'c' } });
-        st.deepEqual(qs.parse('a[]=b&a[hasOwnProperty]=c&a[x]=y', { allowPrototypes: false }), { a: { 0: 'b', x: 'y' } });
-        st.deepEqual(qs.parse('a[]=b&a[hasOwnProperty]=c&a[x]=y', { allowPrototypes: true }), { a: { 0: 'b', hasOwnProperty: 'c', x: 'y' } });
-        st.end();
+        expect(Qs.parse('a[12b]=c')).to.deep.equal({ a: { '12b': 'c' } });
+        done();
     });
 
-    t.test('transforms arrays to objects (dot notation)', function (st) {
-        st.deepEqual(qs.parse('foo[0].baz=bar&fool.bad=baz', { allowDots: true }), { foo: [{ baz: 'bar' }], fool: { bad: 'baz' } });
-        st.deepEqual(qs.parse('foo[0].baz=bar&fool.bad.boo=baz', { allowDots: true }), { foo: [{ baz: 'bar' }], fool: { bad: { boo: 'baz' } } });
-        st.deepEqual(qs.parse('foo[0][0].baz=bar&fool.bad=baz', { allowDots: true }), { foo: [[{ baz: 'bar' }]], fool: { bad: 'baz' } });
-        st.deepEqual(qs.parse('foo[0].baz[0]=15&foo[0].bar=2', { allowDots: true }), { foo: [{ baz: ['15'], bar: '2' }] });
-        st.deepEqual(qs.parse('foo[0].baz[0]=15&foo[0].baz[1]=16&foo[0].bar=2', { allowDots: true }), { foo: [{ baz: ['15', '16'], bar: '2' }] });
-        st.deepEqual(qs.parse('foo.bad=baz&foo[0]=bar', { allowDots: true }), { foo: { bad: 'baz', 0: 'bar' } });
-        st.deepEqual(qs.parse('foo.bad=baz&foo[]=bar', { allowDots: true }), { foo: { bad: 'baz', 0: 'bar' } });
-        st.deepEqual(qs.parse('foo[]=bar&foo.bad=baz', { allowDots: true }), { foo: { 0: 'bar', bad: 'baz' } });
-        st.deepEqual(qs.parse('foo.bad=baz&foo[]=bar&foo[]=foo', { allowDots: true }), { foo: { bad: 'baz', 0: 'bar', 1: 'foo' } });
-        st.deepEqual(qs.parse('foo[0].a=a&foo[0].b=b&foo[1].a=aa&foo[1].b=bb', { allowDots: true }), { foo: [{ a: 'a', b: 'b' }, { a: 'aa', b: 'bb' }] });
-        st.end();
+    it('supports encoded = signs', function (done) {
+
+        expect(Qs.parse('he%3Dllo=th%3Dere')).to.deep.equal({ 'he=llo': 'th=ere' });
+        done();
     });
 
-    t.test('correctly prunes undefined values when converting an array to an object', function (st) {
-        st.deepEqual(qs.parse('a[2]=b&a[99999999]=c'), { a: { 2: 'b', 99999999: 'c' } });
-        st.end();
+    it('is ok with url encoded strings', function (done) {
+
+        expect(Qs.parse('a[b%20c]=d')).to.deep.equal({ a: { 'b c': 'd' } });
+        expect(Qs.parse('a[b]=c%20d')).to.deep.equal({ a: { b: 'c d' } });
+        done();
     });
 
-    t.test('supports malformed uri characters', function (st) {
-        st.deepEqual(qs.parse('{%:%}', { strictNullHandling: true }), { '{%:%}': null });
-        st.deepEqual(qs.parse('{%:%}='), { '{%:%}': '' });
-        st.deepEqual(qs.parse('foo=%:%}'), { foo: '%:%}' });
-        st.end();
+    it('allows brackets in the value', function (done) {
+
+        expect(Qs.parse('pets=["tobi"]')).to.deep.equal({ pets: '["tobi"]' });
+        expect(Qs.parse('operators=[">=", "<="]')).to.deep.equal({ operators: '[">=", "<="]' });
+        done();
     });
 
-    t.test('doesn\'t produce empty keys', function (st) {
-        st.deepEqual(qs.parse('_r=1&'), { _r: '1' });
-        st.end();
+    it('allows empty values', function (done) {
+
+        expect(Qs.parse('')).to.deep.equal({});
+        expect(Qs.parse(null)).to.deep.equal({});
+        expect(Qs.parse(undefined)).to.deep.equal({});
+        done();
     });
 
-    t.test('cannot access Object prototype', function (st) {
-        qs.parse('constructor[prototype][bad]=bad');
-        qs.parse('bad[constructor][prototype][bad]=bad');
-        st.equal(typeof Object.prototype.bad, 'undefined');
-        st.end();
+    it('transforms arrays to objects', function (done) {
+
+        expect(Qs.parse('foo[0]=bar&foo[bad]=baz')).to.deep.equal({ foo: { '0': 'bar', bad: 'baz' } });
+        expect(Qs.parse('foo[bad]=baz&foo[0]=bar')).to.deep.equal({ foo: { bad: 'baz', '0': 'bar' } });
+        expect(Qs.parse('foo[bad]=baz&foo[]=bar')).to.deep.equal({ foo: { bad: 'baz', '0': 'bar' } });
+        expect(Qs.parse('foo[]=bar&foo[bad]=baz')).to.deep.equal({ foo: { '0': 'bar', bad: 'baz' } });
+        expect(Qs.parse('foo[bad]=baz&foo[]=bar&foo[]=foo')).to.deep.equal({ foo: { bad: 'baz', '0': 'bar', '1': 'foo' } });
+        expect(Qs.parse('foo[0][a]=a&foo[0][b]=b&foo[1][a]=aa&foo[1][b]=bb')).to.deep.equal({ foo: [{ a: 'a', b: 'b' }, { a: 'aa', b: 'bb' }] });
+        expect(Qs.parse('a[]=b&a[t]=u&a[hasOwnProperty]=c')).to.deep.equal({ a: { '0': 'b', t: 'u', c: true } });
+        expect(Qs.parse('a[]=b&a[hasOwnProperty]=c&a[x]=y')).to.deep.equal({ a: { '0': 'b', '1': 'c', x: 'y' } });
+        done();
     });
 
-    t.test('parses arrays of objects', function (st) {
-        st.deepEqual(qs.parse('a[][b]=c'), { a: [{ b: 'c' }] });
-        st.deepEqual(qs.parse('a[0][b]=c'), { a: [{ b: 'c' }] });
-        st.end();
+    it('transforms arrays to objects (dot notation)', function (done) {
+
+        expect(Qs.parse('foo[0].baz=bar&fool.bad=baz')).to.deep.equal({ foo: [{ baz: 'bar' }], fool: { bad: 'baz' } });
+        expect(Qs.parse('foo[0].baz=bar&fool.bad.boo=baz')).to.deep.equal({ foo: [{ baz: 'bar' }], fool: { bad: { boo: 'baz' } } });
+        expect(Qs.parse('foo[0][0].baz=bar&fool.bad=baz')).to.deep.equal({ foo: [[{ baz: 'bar' }]], fool: { bad: 'baz' } });
+        expect(Qs.parse('foo[0].baz[0]=15&foo[0].bar=2')).to.deep.equal({ foo: [{ baz: ['15'], bar: '2' }] });
+        expect(Qs.parse('foo[0].baz[0]=15&foo[0].baz[1]=16&foo[0].bar=2')).to.deep.equal({ foo: [{ baz: ['15', '16'], bar: '2' }] });
+        expect(Qs.parse('foo.bad=baz&foo[0]=bar')).to.deep.equal({ foo: { bad: 'baz', '0': 'bar' } });
+        expect(Qs.parse('foo.bad=baz&foo[]=bar')).to.deep.equal({ foo: { bad: 'baz', '0': 'bar' } });
+        expect(Qs.parse('foo[]=bar&foo.bad=baz')).to.deep.equal({ foo: { '0': 'bar', bad: 'baz' } });
+        expect(Qs.parse('foo.bad=baz&foo[]=bar&foo[]=foo')).to.deep.equal({ foo: { bad: 'baz', '0': 'bar', '1': 'foo' } });
+        expect(Qs.parse('foo[0].a=a&foo[0].b=b&foo[1].a=aa&foo[1].b=bb')).to.deep.equal({ foo: [{ a: 'a', b: 'b' }, { a: 'aa', b: 'bb' }] });
+        done();
     });
 
-    t.test('allows for empty strings in arrays', function (st) {
-        st.deepEqual(qs.parse('a[]=b&a[]=&a[]=c'), { a: ['b', '', 'c'] });
+    it('can add keys to objects', function (done) {
 
-        st.deepEqual(
-            qs.parse('a[0]=b&a[1]&a[2]=c&a[19]=', { strictNullHandling: true, arrayLimit: 20 }),
-            { a: ['b', null, 'c', ''] },
-            'with arrayLimit 20 + array indices: null then empty string works'
-        );
-        st.deepEqual(
-            qs.parse('a[]=b&a[]&a[]=c&a[]=', { strictNullHandling: true, arrayLimit: 0 }),
-            { a: ['b', null, 'c', ''] },
-            'with arrayLimit 0 + array brackets: null then empty string works'
-        );
-
-        st.deepEqual(
-            qs.parse('a[0]=b&a[1]=&a[2]=c&a[19]', { strictNullHandling: true, arrayLimit: 20 }),
-            { a: ['b', '', 'c', null] },
-            'with arrayLimit 20 + array indices: empty string then null works'
-        );
-        st.deepEqual(
-            qs.parse('a[]=b&a[]=&a[]=c&a[]', { strictNullHandling: true, arrayLimit: 0 }),
-            { a: ['b', '', 'c', null] },
-            'with arrayLimit 0 + array brackets: empty string then null works'
-        );
-
-        st.deepEqual(
-            qs.parse('a[]=&a[]=b&a[]=c'),
-            { a: ['', 'b', 'c'] },
-            'array brackets: empty strings work'
-        );
-        st.end();
+        expect(Qs.parse('a[b]=c&a=d')).to.deep.equal({ a: { b: 'c', d: true } });
+        done();
     });
 
-    t.test('compacts sparse arrays', function (st) {
-        st.deepEqual(qs.parse('a[10]=1&a[2]=2', { arrayLimit: 20 }), { a: ['2', '1'] });
-        st.deepEqual(qs.parse('a[1][b][2][c]=1', { arrayLimit: 20 }), { a: [{ b: [{ c: '1' }] }] });
-        st.deepEqual(qs.parse('a[1][2][3][c]=1', { arrayLimit: 20 }), { a: [[[{ c: '1' }]]] });
-        st.deepEqual(qs.parse('a[1][2][3][c][1]=1', { arrayLimit: 20 }), { a: [[[{ c: ['1'] }]]] });
-        st.end();
+    it('correctly prunes undefined values when converting an array to an object', function (done) {
+
+        expect(Qs.parse('a[2]=b&a[99999999]=c')).to.deep.equal({ a: { '2': 'b', '99999999': 'c' } });
+        done();
     });
 
-    t.test('parses semi-parsed strings', function (st) {
-        st.deepEqual(qs.parse({ 'a[b]': 'c' }), { a: { b: 'c' } });
-        st.deepEqual(qs.parse({ 'a[b]': 'c', 'a[d]': 'e' }), { a: { b: 'c', d: 'e' } });
-        st.end();
+    it('supports malformed uri characters', function (done) {
+
+        expect(Qs.parse('{%:%}', { strictNullHandling: true })).to.deep.equal({ '{%:%}': null });
+        expect(Qs.parse('{%:%}=')).to.deep.equal({ '{%:%}': '' });
+        expect(Qs.parse('foo=%:%}')).to.deep.equal({ foo: '%:%}' });
+        done();
     });
 
-    t.test('parses buffers correctly', function (st) {
+    it('doesn\'t produce empty keys', function (done) {
+
+        expect(Qs.parse('_r=1&')).to.deep.equal({ '_r': '1' });
+        done();
+    });
+
+    it('cannot access Object prototype', function (done) {
+
+        Qs.parse('constructor[prototype][bad]=bad');
+        Qs.parse('bad[constructor][prototype][bad]=bad');
+        expect(typeof Object.prototype.bad).to.equal('undefined');
+        done();
+    });
+
+    it('parses arrays of objects', function (done) {
+
+        expect(Qs.parse('a[][b]=c')).to.deep.equal({ a: [{ b: 'c' }] });
+        expect(Qs.parse('a[0][b]=c')).to.deep.equal({ a: [{ b: 'c' }] });
+        done();
+    });
+
+    it('allows for empty strings in arrays', function (done) {
+
+        expect(Qs.parse('a[]=b&a[]=&a[]=c')).to.deep.equal({ a: ['b', '', 'c'] });
+        expect(Qs.parse('a[0]=b&a[1]&a[2]=c&a[19]=', { strictNullHandling: true })).to.deep.equal({ a: ['b', null, 'c', ''] });
+        expect(Qs.parse('a[0]=b&a[1]=&a[2]=c&a[19]', { strictNullHandling: true })).to.deep.equal({ a: ['b', '', 'c', null] });
+        expect(Qs.parse('a[]=&a[]=b&a[]=c')).to.deep.equal({ a: ['', 'b', 'c'] });
+        done();
+    });
+
+    it('compacts sparse arrays', function (done) {
+
+        expect(Qs.parse('a[10]=1&a[2]=2')).to.deep.equal({ a: ['2', '1'] });
+        done();
+    });
+
+    it('parses semi-parsed strings', function (done) {
+
+        expect(Qs.parse({ 'a[b]': 'c' })).to.deep.equal({ a: { b: 'c' } });
+        expect(Qs.parse({ 'a[b]': 'c', 'a[d]': 'e' })).to.deep.equal({ a: { b: 'c', d: 'e' } });
+        done();
+    });
+
+    it('parses buffers correctly', function (done) {
+
         var b = new Buffer('test');
-        st.deepEqual(qs.parse({ a: b }), { a: b });
-        st.end();
+        expect(Qs.parse({ a: b })).to.deep.equal({ a: b });
+        done();
     });
 
-    t.test('continues parsing when no parent is found', function (st) {
-        st.deepEqual(qs.parse('[]=&a=b'), { 0: '', a: 'b' });
-        st.deepEqual(qs.parse('[]&a=b', { strictNullHandling: true }), { 0: null, a: 'b' });
-        st.deepEqual(qs.parse('[foo]=bar'), { foo: 'bar' });
-        st.end();
+    it('continues parsing when no parent is found', function (done) {
+
+        expect(Qs.parse('[]=&a=b')).to.deep.equal({ '0': '', a: 'b' });
+        expect(Qs.parse('[]&a=b', { strictNullHandling: true })).to.deep.equal({ '0': null, a: 'b' });
+        expect(Qs.parse('[foo]=bar')).to.deep.equal({ foo: 'bar' });
+        done();
     });
 
-    t.test('does not error when parsing a very long array', function (st) {
+    it('does not error when parsing a very long array', function (done) {
+
         var str = 'a[]=a';
         while (Buffer.byteLength(str) < 128 * 1024) {
-            str = str + '&' + str;
+            str += '&' + str;
         }
 
-        st.doesNotThrow(function () {
-            qs.parse(str);
-        });
+        expect(function () {
 
-        st.end();
+            Qs.parse(str);
+        }).to.not.throw();
+
+        done();
     });
 
-    t.test('should not throw when a native prototype has an enumerable property', { parallel: false }, function (st) {
+    it('should not throw when a native prototype has an enumerable property', { parallel: false }, function (done) {
+
         Object.prototype.crash = '';
         Array.prototype.crash = '';
-        st.doesNotThrow(qs.parse.bind(null, 'a=b'));
-        st.deepEqual(qs.parse('a=b'), { a: 'b' });
-        st.doesNotThrow(qs.parse.bind(null, 'a[][b]=c'));
-        st.deepEqual(qs.parse('a[][b]=c'), { a: [{ b: 'c' }] });
+        expect(Qs.parse.bind(null, 'a=b')).to.not.throw();
+        expect(Qs.parse('a=b')).to.deep.equal({ a: 'b' });
+        expect(Qs.parse.bind(null, 'a[][b]=c')).to.not.throw();
+        expect(Qs.parse('a[][b]=c')).to.deep.equal({ a: [{ b: 'c' }] });
         delete Object.prototype.crash;
         delete Array.prototype.crash;
-        st.end();
+        done();
     });
 
-    t.test('parses a string with an alternative string delimiter', function (st) {
-        st.deepEqual(qs.parse('a=b;c=d', { delimiter: ';' }), { a: 'b', c: 'd' });
-        st.end();
+    it('parses a string with an alternative string delimiter', function (done) {
+
+        expect(Qs.parse('a=b;c=d', { delimiter: ';' })).to.deep.equal({ a: 'b', c: 'd' });
+        done();
     });
 
-    t.test('parses a string with an alternative RegExp delimiter', function (st) {
-        st.deepEqual(qs.parse('a=b; c=d', { delimiter: /[;,] */ }), { a: 'b', c: 'd' });
-        st.end();
+    it('parses a string with an alternative RegExp delimiter', function (done) {
+
+        expect(Qs.parse('a=b; c=d', { delimiter: /[;,] */ })).to.deep.equal({ a: 'b', c: 'd' });
+        done();
     });
 
-    t.test('does not use non-splittable objects as delimiters', function (st) {
-        st.deepEqual(qs.parse('a=b&c=d', { delimiter: true }), { a: 'b', c: 'd' });
-        st.end();
+    it('does not use non-splittable objects as delimiters', function (done) {
+
+        expect(Qs.parse('a=b&c=d', { delimiter: true })).to.deep.equal({ a: 'b', c: 'd' });
+        done();
     });
 
-    t.test('allows overriding parameter limit', function (st) {
-        st.deepEqual(qs.parse('a=b&c=d', { parameterLimit: 1 }), { a: 'b' });
-        st.end();
+    it('allows overriding parameter limit', function (done) {
+
+        expect(Qs.parse('a=b&c=d', { parameterLimit: 1 })).to.deep.equal({ a: 'b' });
+        done();
     });
 
-    t.test('allows setting the parameter limit to Infinity', function (st) {
-        st.deepEqual(qs.parse('a=b&c=d', { parameterLimit: Infinity }), { a: 'b', c: 'd' });
-        st.end();
+    it('allows setting the parameter limit to Infinity', function (done) {
+
+        expect(Qs.parse('a=b&c=d', { parameterLimit: Infinity })).to.deep.equal({ a: 'b', c: 'd' });
+        done();
     });
 
-    t.test('allows overriding array limit', function (st) {
-        st.deepEqual(qs.parse('a[0]=b', { arrayLimit: -1 }), { a: { 0: 'b' } });
-        st.deepEqual(qs.parse('a[-1]=b', { arrayLimit: -1 }), { a: { '-1': 'b' } });
-        st.deepEqual(qs.parse('a[0]=b&a[1]=c', { arrayLimit: 0 }), { a: { 0: 'b', 1: 'c' } });
-        st.end();
+    it('allows overriding array limit', function (done) {
+
+        expect(Qs.parse('a[0]=b', { arrayLimit: -1 })).to.deep.equal({ a: { '0': 'b' } });
+        expect(Qs.parse('a[-1]=b', { arrayLimit: -1 })).to.deep.equal({ a: { '-1': 'b' } });
+        expect(Qs.parse('a[0]=b&a[1]=c', { arrayLimit: 0 })).to.deep.equal({ a: { '0': 'b', '1': 'c' } });
+        done();
     });
 
-    t.test('allows disabling array parsing', function (st) {
-        st.deepEqual(qs.parse('a[0]=b&a[1]=c', { parseArrays: false }), { a: { 0: 'b', 1: 'c' } });
-        st.end();
+    it('allows disabling array parsing', function (done) {
+
+        expect(Qs.parse('a[0]=b&a[1]=c', { parseArrays: false })).to.deep.equal({ a: { '0': 'b', '1': 'c' } });
+        done();
     });
 
-    t.test('parses an object', function (st) {
+    it('parses an object', function (done) {
+
         var input = {
             'user[name]': { 'pop[bob]': 3 },
             'user[email]': null
         };
 
         var expected = {
-            user: {
-                name: { 'pop[bob]': 3 },
-                email: null
+            'user': {
+                'name': { 'pop[bob]': 3 },
+                'email': null
             }
         };
 
-        var result = qs.parse(input);
+        var result = Qs.parse(input);
 
-        st.deepEqual(result, expected);
-        st.end();
+        expect(result).to.deep.equal(expected);
+        done();
     });
 
-    t.test('parses an object in dot notation', function (st) {
+    it('parses an object in dot notation', function (done) {
+
         var input = {
             'user.name': { 'pop[bob]': 3 },
             'user.email.': null
         };
 
         var expected = {
-            user: {
-                name: { 'pop[bob]': 3 },
-                email: null
+            'user': {
+                'name': { 'pop[bob]': 3 },
+                'email': null
             }
         };
 
-        var result = qs.parse(input, { allowDots: true });
+        var result = Qs.parse(input);
 
-        st.deepEqual(result, expected);
-        st.end();
+        expect(result).to.deep.equal(expected);
+        done();
     });
 
-    t.test('parses an object and not child values', function (st) {
+    it('parses an object and not child values', function (done) {
+
         var input = {
-            'user[name]': { 'pop[bob]': { test: 3 } },
+            'user[name]': { 'pop[bob]': { 'test': 3 } },
             'user[email]': null
         };
 
         var expected = {
-            user: {
-                name: { 'pop[bob]': { test: 3 } },
-                email: null
+            'user': {
+                'name': { 'pop[bob]': { 'test': 3 } },
+                'email': null
             }
         };
 
-        var result = qs.parse(input);
+        var result = Qs.parse(input);
 
-        st.deepEqual(result, expected);
-        st.end();
+        expect(result).to.deep.equal(expected);
+        done();
     });
 
-    t.test('does not blow up when Buffer global is missing', function (st) {
+    it('does not blow up when Buffer global is missing', function (done) {
+
         var tempBuffer = global.Buffer;
         delete global.Buffer;
-        var result = qs.parse('a=b&c=d');
+        var result = Qs.parse('a=b&c=d');
         global.Buffer = tempBuffer;
-        st.deepEqual(result, { a: 'b', c: 'd' });
-        st.end();
+        expect(result).to.deep.equal({ a: 'b', c: 'd' });
+        done();
     });
 
-    t.test('does not crash when parsing circular references', function (st) {
+    it('does not crash when parsing circular references', function (done) {
+
         var a = {};
         a.b = a;
 
         var parsed;
 
-        st.doesNotThrow(function () {
-            parsed = qs.parse({ 'foo[bar]': 'baz', 'foo[baz]': a });
-        });
+        expect(function () {
 
-        st.equal('foo' in parsed, true, 'parsed has "foo" property');
-        st.equal('bar' in parsed.foo, true);
-        st.equal('baz' in parsed.foo, true);
-        st.equal(parsed.foo.bar, 'baz');
-        st.deepEqual(parsed.foo.baz, a);
-        st.end();
+            parsed = Qs.parse({ 'foo[bar]': 'baz', 'foo[baz]': a });
+        }).to.not.throw();
+
+        expect(parsed).to.contain('foo');
+        expect(parsed.foo).to.contain('bar', 'baz');
+        expect(parsed.foo.bar).to.equal('baz');
+        expect(parsed.foo.baz).to.deep.equal(a);
+        done();
     });
 
-    t.test('parses null objects correctly', { skip: !Object.create }, function (st) {
+    it('parses plain objects correctly', function (done) {
+
         var a = Object.create(null);
         a.b = 'c';
 
-        st.deepEqual(qs.parse(a), { b: 'c' });
-        var result = qs.parse({ a: a });
-        st.equal('a' in result, true, 'result has "a" property');
-        st.deepEqual(result.a, a);
-        st.end();
+        expect(Qs.parse(a)).to.deep.equal({ b: 'c' });
+        var result = Qs.parse({ a: a });
+        expect(result).to.contain('a');
+        expect(result.a).to.deep.equal(a);
+        done();
     });
 
-    t.test('parses dates correctly', function (st) {
+    it('parses dates correctly', function (done) {
+
         var now = new Date();
-        st.deepEqual(qs.parse({ a: now }), { a: now });
-        st.end();
+        expect(Qs.parse({ a: now })).to.deep.equal({ a: now });
+        done();
     });
 
-    t.test('parses regular expressions correctly', function (st) {
+    it('parses regular expressions correctly', function (done) {
+
         var re = /^test$/;
-        st.deepEqual(qs.parse({ a: re }), { a: re });
-        st.end();
+        expect(Qs.parse({ a: re })).to.deep.equal({ a: re });
+        done();
     });
 
-    t.test('does not allow overwriting prototype properties', function (st) {
-        st.deepEqual(qs.parse('a[hasOwnProperty]=b', { allowPrototypes: false }), {});
-        st.deepEqual(qs.parse('hasOwnProperty=b', { allowPrototypes: false }), {});
+    it('can allow overwriting prototype properties', function (done) {
 
-        st.deepEqual(
-            qs.parse('toString', { allowPrototypes: false }),
-            {},
-            'bare "toString" results in {}'
-        );
-
-        st.end();
+        expect(Qs.parse('a[hasOwnProperty]=b', { allowPrototypes: true })).to.deep.equal({ a: { hasOwnProperty: 'b' } }, { prototype: false });
+        expect(Qs.parse('hasOwnProperty=b', { allowPrototypes: true })).to.deep.equal({ hasOwnProperty: 'b' }, { prototype: false });
+        done();
     });
 
-    t.test('can allow overwriting prototype properties', function (st) {
-        st.deepEqual(qs.parse('a[hasOwnProperty]=b', { allowPrototypes: true }), { a: { hasOwnProperty: 'b' } });
-        st.deepEqual(qs.parse('hasOwnProperty=b', { allowPrototypes: true }), { hasOwnProperty: 'b' });
+    it('can return plain objects', function (done) {
 
-        st.deepEqual(
-            qs.parse('toString', { allowPrototypes: true }),
-            { toString: '' },
-            'bare "toString" results in { toString: "" }'
-        );
-
-        st.end();
-    });
-
-    t.test('params starting with a closing bracket', function (st) {
-        st.deepEqual(qs.parse(']=toString'), { ']': 'toString' });
-        st.deepEqual(qs.parse(']]=toString'), { ']]': 'toString' });
-        st.deepEqual(qs.parse(']hello]=toString'), { ']hello]': 'toString' });
-        st.end();
-    });
-
-    t.test('params starting with a starting bracket', function (st) {
-        st.deepEqual(qs.parse('[=toString'), { '[': 'toString' });
-        st.deepEqual(qs.parse('[[=toString'), { '[[': 'toString' });
-        st.deepEqual(qs.parse('[hello[=toString'), { '[hello[': 'toString' });
-        st.end();
-    });
-
-    t.test('add keys to objects', function (st) {
-        st.deepEqual(
-            qs.parse('a[b]=c&a=d'),
-            { a: { b: 'c', d: true } },
-            'can add keys to objects'
-        );
-
-        st.deepEqual(
-            qs.parse('a[b]=c&a=toString'),
-            { a: { b: 'c' } },
-            'can not overwrite prototype'
-        );
-
-        st.deepEqual(
-            qs.parse('a[b]=c&a=toString', { allowPrototypes: true }),
-            { a: { b: 'c', toString: true } },
-            'can overwrite prototype with allowPrototypes true'
-        );
-
-        st.deepEqual(
-            qs.parse('a[b]=c&a=toString', { plainObjects: true }),
-            { a: { b: 'c', toString: true } },
-            'can overwrite prototype with plainObjects true'
-        );
-
-        st.end();
-    });
-
-    t.test('can return null objects', { skip: !Object.create }, function (st) {
         var expected = Object.create(null);
         expected.a = Object.create(null);
         expected.a.b = 'c';
         expected.a.hasOwnProperty = 'd';
-        st.deepEqual(qs.parse('a[b]=c&a[hasOwnProperty]=d', { plainObjects: true }), expected);
-        st.deepEqual(qs.parse(null, { plainObjects: true }), Object.create(null));
+        expect(Qs.parse('a[b]=c&a[hasOwnProperty]=d', { plainObjects: true })).to.deep.equal(expected);
+        expect(Qs.parse(null, { plainObjects: true })).to.deep.equal(Object.create(null));
         var expectedArray = Object.create(null);
         expectedArray.a = Object.create(null);
-        expectedArray.a[0] = 'b';
+        expectedArray.a['0'] = 'b';
         expectedArray.a.c = 'd';
-        st.deepEqual(qs.parse('a[]=b&a[c]=d', { plainObjects: true }), expectedArray);
-        st.end();
-    });
-
-    t.test('can parse with custom encoding', function (st) {
-        st.deepEqual(qs.parse('%8c%a7=%91%e5%8d%e3%95%7b', {
-            decoder: function (str) {
-                var reg = /%([0-9A-F]{2})/ig;
-                var result = [];
-                var parts = reg.exec(str);
-                while (parts) {
-                    result.push(parseInt(parts[1], 16));
-                    parts = reg.exec(str);
-                }
-                return iconv.decode(new Buffer(result), 'shift_jis').toString();
-            }
-        }), { 県: '大阪府' });
-        st.end();
-    });
-
-    t.test('throws error with wrong decoder', function (st) {
-        st.throws(function () {
-            qs.parse({}, { decoder: 'string' });
-        }, new TypeError('Decoder has to be a function.'));
-        st.end();
+        expect(Qs.parse('a[]=b&a[c]=d', { plainObjects: true })).to.deep.equal(expectedArray);
+        done();
     });
 });
