@@ -88,7 +88,6 @@ var CreatableSelect = function (_React$Component) {
 			var _props2 = this.props,
 			    filterOptions = _props2.filterOptions,
 			    isValidNewOption = _props2.isValidNewOption,
-			    options = _props2.options,
 			    promptTextCreator = _props2.promptTextCreator;
 
 			// TRICKY Check currently selected options as well.
@@ -302,9 +301,9 @@ function shouldKeyDownEventCreateNewOption(_ref6) {
 		case 188:
 			// COMMA
 			return true;
+		default:
+			return false;
 	}
-
-	return false;
 };
 
 // Default prop methods
