@@ -143,7 +143,19 @@ var Select = function (_React$Component) {
 			if (this.menu && this.focused && this.state.isOpen && !this.hasScrolledToOption) {
 				var focusedOptionNode = (0, _reactDom.findDOMNode)(this.focused);
 				var menuNode = (0, _reactDom.findDOMNode)(this.menu);
-				menuNode.scrollTop = focusedOptionNode.offsetTop;
+
+				var scrollTop = menuNode.scrollTop;
+				var scrollBottom = scrollTop + menuNode.offsetHeight;
+				var optionTop = focusedOptionNode.offsetTop;
+				var optionBottom = optionTop + focusedOptionNode.offsetHeight;
+
+				if (scrollTop > optionTop || scrollBottom < optionBottom) {
+					menuNode.scrollTop = focusedOptionNode.offsetTop;
+				}
+
+				// We still set hasScrolledToOption to true even if we didn't
+				// actually need to scroll, as we've still confirmed that the
+				// option is in view.
 				this.hasScrolledToOption = true;
 			} else if (!this.state.isOpen) {
 				this.hasScrolledToOption = false;
@@ -261,11 +273,19 @@ var Select = function (_React$Component) {
 			}
 
 			if (event.target.tagName === 'INPUT') {
+				if (!this.state.isFocused) {
+					this._openAfterFocus = this.props.openOnClick;
+					this.focus();
+				} else if (!this.state.isOpen) {
+					this.setState({
+						isOpen: true,
+						isPseudoFocused: false
+					});
+				}
 				return;
 			}
 
 			// prevent default event handlers
-			event.stopPropagation();
 			event.preventDefault();
 
 			// for the non-searchable select, toggle the menu
@@ -313,13 +333,17 @@ var Select = function (_React$Component) {
 			}
 			// If the menu isn't open, let the event bubble to the main handleMouseDown
 			if (!this.state.isOpen) {
-				return;
+				this.setState({
+					isOpen: true
+				});
 			}
 			// prevent default event handlers
 			event.stopPropagation();
 			event.preventDefault();
 			// close the menu
-			this.closeMenu();
+			if (this.state.isOpen) {
+				this.closeMenu();
+			}
 		}
 	}, {
 		key: 'handleMouseDownOnMenu',
@@ -699,7 +723,6 @@ var Select = function (_React$Component) {
 			if (event && event.type === 'mousedown' && event.button !== 0) {
 				return;
 			}
-			event.stopPropagation();
 			event.preventDefault();
 			this.setValue(this.getResetValue());
 			this.setState({
