@@ -16,17 +16,21 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 function menuRenderer(_ref) {
 	var focusedOption = _ref.focusedOption,
+	    focusOption = _ref.focusOption,
+	    inputValue = _ref.inputValue,
 	    instancePrefix = _ref.instancePrefix,
 	    labelKey = _ref.labelKey,
 	    onFocus = _ref.onFocus,
+	    onOptionRef = _ref.onOptionRef,
 	    onSelect = _ref.onSelect,
 	    optionClassName = _ref.optionClassName,
 	    optionComponent = _ref.optionComponent,
 	    optionRenderer = _ref.optionRenderer,
 	    options = _ref.options,
+	    removeValue = _ref.removeValue,
+	    selectValue = _ref.selectValue,
 	    valueArray = _ref.valueArray,
-	    valueKey = _ref.valueKey,
-	    onOptionRef = _ref.onOptionRef;
+	    valueKey = _ref.valueKey;
 
 	var Option = optionComponent;
 
@@ -46,6 +50,8 @@ function menuRenderer(_ref) {
 			Option,
 			{
 				className: optionClass,
+				focusOption: focusOption,
+				inputValue: inputValue,
 				instancePrefix: instancePrefix,
 				isDisabled: option.disabled,
 				isFocused: isFocused,
@@ -57,9 +63,11 @@ function menuRenderer(_ref) {
 				optionIndex: i,
 				ref: function ref(_ref2) {
 					onOptionRef(_ref2, isFocused);
-				}
+				},
+				removeValue: removeValue,
+				selectValue: selectValue
 			},
-			optionRenderer(option, i)
+			optionRenderer(option, i, inputValue)
 		);
 	});
 }
