@@ -323,7 +323,9 @@ function Server(compiler, options) {
     },
 
     before: () => {
-      if (typeof options.before === 'function') { options.before(app, this); }
+      if (typeof options.before === 'function') {
+        options.before(app, this);
+      }
     },
 
     middleware: () => {
